requirements.txt
django==3.0a1
pyyaml
exos
